tidy + readme update
