+++ src/getters.js
@@ -1,0 +1,5 @@
+export const getters = {
+    getPage: (state) => {
+        return state.currentPage
+    }
+}
