+++ src/mutations.js
@@ -1,0 +1,5 @@
+export const mutations = {
+    updateCurrentPage(state, n) {
+      state.currentPage = n
+    }
+  }
